+++ app/src/main/java/shaun/home/MainActivity.java
@@ -26,13 +26,16 @@ import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
 import android.graphics.drawable.Drawable;
 import android.os.PowerManager;
+import android.provider.Settings;
 import android.support.v4.app.FragmentStatePagerAdapter;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.GestureDetector;
 import android.view.GestureDetector.OnGestureListener;
 import android.view.LayoutInflater;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.Window;
 import android.widget.ImageButton;
@@ -87,7 +90,6 @@ public class MainActivity extends AppCompatActivity{
         }
         @Override
         public void onClick(View v) {
-            lock();
             RelativeLayout main = ((RelativeLayout) findViewById(R.id.main_activity));
             int temp = currentTapped;
             if(currentTapped!=0){
@@ -146,45 +148,4 @@ public class MainActivity extends AppCompatActivity{
         return resolveInfo.loadIcon(pm);
     }
 
-
-    public class AdminReceiver extends DeviceAdminReceiver {
-        public static final String ACTION_DISABLED = "device_admin_action_disabled";
-        public static final String ACTION_ENABLED = "device_admin_action_enabled";
-
-        public AdminReceiver(){
-            super();
-        }
-        @Override
-        public void onDisabled(Context context, Intent intent) {
-            super.onDisabled(context, intent);
-            LocalBroadcastManager.getInstance(context).sendBroadcast(
-                    new Intent(ACTION_DISABLED));
-        }
-        @Override
-        public void onEnabled(Context context, Intent intent) {
-            super.onEnabled(context, intent);
-            LocalBroadcastManager.getInstance(context).sendBroadcast(
-                    new Intent(ACTION_ENABLED));
-        }
-    }
-    private void lock() {
-        PowerManager pm = (PowerManager)getSystemService(Context.POWER_SERVICE);
-        if (pm.isScreenOn()) {
-            DevicePolicyManager policy = (DevicePolicyManager)
-                    getSystemService(Context.DEVICE_POLICY_SERVICE);
-            try {
-                policy.lockNow();
-            } catch (SecurityException ex) {
-                Toast.makeText(
-                        this,
-                        "must enable device administrator",
-                        Toast.LENGTH_LONG).show();
-                ComponentName admin = new ComponentName(this, AdminReceiver.class);
-                Intent intent = new Intent(
-                        DevicePolicyManager.ACTION_ADD_DEVICE_ADMIN).putExtra(
-                        DevicePolicyManager.EXTRA_DEVICE_ADMIN, admin);
-                this.startActivity(intent);
-            }
-        }
-    }
 }
