+++ app/src/main/res/xml/device_admin_sample.xml
@@ -1,6 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <device-admin xmlns:android="http://schemas.android.com/apk/res/android">
     <uses-policies>
+        <limit-password />
+        <watch-login />
+        <reset-password />
         <force-lock />
+        <wipe-data />
     </uses-policies>
 </device-admin>
